requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1